+++ lib/xdf_writer_template.h
@@ -5,6 +5,6 @@
 #include "config.h"
 void write_header(XDFWriter* writer, const Config& cfg );
 
-void write_footer(XDFWriter* writer,const Config& cfg ,double exact_ts, double time_stamp);
+void write_footer(XDFWriter* writer,const Config& cfg ,double exact_ts, long time_stamp);
 
 #endif //XDF_WRITER_TEMPLATE_H
